v3.0.1-dev Changes for 3.0 p2
[Settings]
* Replaced str_cleanLevel with: (resolves #11)
  * cln_1enable_bool, cln_aurpkg_bool, cln_aurbuild_bool, cln_paccache_num
* Added cln_orphan_bool
* Renamed/reorganized most settings (resolves #10)

[Functionality]
* Old settings are converted to new (new keep existing settings)
* Added basic config validation (falls back to default if invalid)
* Config file is re-generated at start instead of at end

[Aesthetics]
* Reorganized settings according to category
* Made generated config file easier to read
* Silence some output
* Misc code cleanup/reorganization

[ReadMe]
* Updated settings and descriptions
* Reorganized settings based on new format
* Added some additional info

+++ auto-update.sh
@@ -1,105 +1,103 @@
 #!/bin/bash
 #Auto Update For Manjaro Xfce by Lectrode
-vsn="v3.0.0-dev"; vsndsp="$vsn 2018-09-03"
+vsn="v3.0.1-dev"; vsndsp="$vsn 2018-09-13"
 #-Downloads and Installs new updates
 #-Depends: pacman, paccache, xfce4-notifyd, grep, ping
 #-Optional Depends: pikaur, apacman (deprecated)
 true=0; false=1; ctrue=1; cfalse=0;
 
+
+conf_f='/etc/xs/auto-update.conf'
 debgn=+x; # -x =debugging | +x =no debugging
 set $debgn
 
 
-#---Load/Generate config---
-conf_f='/etc/xs/auto-update.conf'
-typeset -A flag_a
-typeset -A conf_a; conf_a=(
-    [aur_1helper_str]="auto"
-    [aur_devel_bool]=$ctrue
-    [notify_lastmsg_num]=20
-    [bool_Downgrades]="$ctrue"
-    [bool_detectErrors]=$ctrue
-    [bool_updateKeys]=$ctrue
-    [bool_updateFlatpak]=$ctrue
-    [bool_notifyMe]=$ctrue
-    [str_ignorePackages]=""
-    [str_mirrorCountry]=""
-    [str_testSite]="www.google.com"
-    [str_cleanLevel]="high" #high, low, off
-    [str_log_d]="/var/log/xs" )
 
-if [ -f $conf_f ]; then
-    while read line; do
-        if ! echo ${line:0:1} | grep -E '#|;' >/dev/null; then
-            if echo $line | grep -F = &>/dev/null; then
-                varname=$(echo "$line" | cut -d '=' -f 1)
-                line=$(echo "$line" | cut -d '=' -f 2-)
-                line=$(echo "$line" | cut -d ';' -f 1 | cut -d '#' -f 1)
-                if [[ "$line" = "" ]]; then echo "$varname" | grep -F "num" >/dev/null && let "line += 0"; fi
-                [[ "$line" = "" ]] || conf_a[$varname]=$line
-                echo "$varname" | grep -F "zflag:" >/dev/null && \
-                flag_a[$(echo "$varname" | cut -d ':' -f 2)]="${conf_a[$varname]}"
-            fi
-        fi
-    done < "$conf_f"; unset line; unset varname
+#---Define Functions---
+
+trouble(){ (echo;echo "#XS# `date` - $@") |tee -a $log_f; }
+troublem(){ echo "XS-$@" |tee -a $log_f; }
+
+pacclean(){
+[[ ! "${conf_a[cln_1enable_bool]}" = "$ctrue" ]] && return
+
+[[ "$(expr ${conf_a[cln_aurpkg_bool]} + ${conf_a[cln_aurbuild_bool]} + ${conf_a[cln_paccache_num]})" -gt "-1" ]] && trouble "Performing cleanup operations..."
+
+if [[ "${conf_a[cln_aurpkg_bool]}" = "$ctrue" ]]; then
+    troublem "Cleaning AUR package cache..."
+    if [ -d /var/cache/apacman/pkg ]; then rm -rf /var/cache/apacman/pkg/*; fi
+    if [ -d /var/cache/pikaur/pkg ]; then rm -rf /var/cache/pikaur/pkg/*; fi
 fi
+
+if [[ "${conf_a[cln_aurbuild_bool]}" = "$ctrue" ]]; then
+    troublem "Cleaning AUR build cache..."
+    if [ -d /var/cache/pikaur/aur_repos ]; then rm -rf /var/cache/pikaur/aur_repos/*; fi
+    if [ -d /var/cache/pikaur/build ]; then rm -rf /var/cache/pikaur/build/*; fi
+fi
+
+if [[ "${conf_a[cln_paccache_num]}" -gt "-1" ]]; then
+    troublem "Cleaning pacman cache..."
+    paccache -rfqk${conf_a[cln_paccache_num]}
+fi
+}
 
 exportconf(){
 if [ ! -d `dirname $conf_f` ]; then mkdir `dirname $conf_f`; fi
-echo '#Config for XS-AutoUpdate' | sudo tee "$conf_f"
-echo '#' | sudo tee -a "$conf_f"
-echo '#NOTES:' | sudo tee -a "$conf_f"
-echo '#aur_1helper_str: Valid options are auto,none,all,pikaur,apacman' | sudo tee -a "$conf_f"
-echo '#aur_devel_bool: Check for AUR devel package updates (requires pikaur AURHelper)' | sudo tee -a "$conf_f"
-echo '#notify_lastmsg_num: Seconds before final normal notification expires (0=never)' | sudo tee -a "$conf_f"
-echo '#bool_Downgrades: Directs pacman to downgrade package if remote is older than local' | sudo tee -a "$conf_f"
-echo '#bool_detectErrors: Include possible errors in notifications' | sudo tee -a "$conf_f"
-echo '#bool_updateKeys: Check for security signature/key updates' | sudo tee -a "$conf_f"
-echo '#bool_updateFlatpak: Check for Flatpak package updates' | sudo tee -a "$conf_f"
-echo '#bool_notifyMe: Enable/Disable nofications' | sudo tee -a "$conf_f"
-echo '#str_cleanLevel: high, low, or off. how much cleaning is done before/after update' | sudo tee -a "$conf_f"
-echo '#str_ignorePackages: list of packages to ignore separated by spaces (in addition to pacman.conf)' | sudo tee -a "$conf_f"
-echo '#str_mirrorCountry: Countries separated by commas from which to pull updates. Default is automatic (geoip)' | sudo tee -a "$conf_f"
-echo '#str_testSite: url (without protocol) used to test internet connection' | sudo tee -a "$conf_f"
-echo '#str_log_d: path to the log directory' | sudo tee -a "$conf_f"
-echo '#' | sudo tee -a "$conf_f"
-echo '#You can also specify makepkg flags for specific AUR packages (requires pikaur AURHelper):' | sudo tee -a "$conf_f"
-echo '#zflag:packagename1,packagename2=--flag1,--flag2,--flag3' | sudo tee -a "$conf_f"
-echo '#' | sudo tee -a "$conf_f"
+echo '#Config for XS-AutoUpdate' > "$conf_f"
+echo '#' >> "$conf_f"
+echo '# AUR Settings #' >> "$conf_f"
+echo '#aur_1helper_str: Valid options are auto,none,all,pikaur,apacman' >> "$conf_f"
+echo '#aur_devel_bool:  If enabled, directs pikaur to update -git and -svn packages' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Cleanup Settings #' >> "$conf_f"
+echo '#cln_1enable_bool:  Enables/disables all package cleanup' >> "$conf_f"
+echo '#cln_aurpkg_bool:   Enables/disables AUR package cleanup' >> "$conf_f"
+echo '#cln_aurbuild_bool: Enables/disables AUR build cleanup' >> "$conf_f"
+echo '#cln_orphan_bool:   Enables/disables uninstall of uneeded packages' >> "$conf_f"
+echo '#cln_paccache_num:  Number of official packages to keep (-1 to keep all)' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Flatpak Settings #' >> "$conf_f"
+echo '#flatpak_1enable_bool: Check for Flatpak package updates' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Notification Settings #' >> "$conf_f"
+echo '#notify_1enable_bool: Enable/Disable nofications' >> "$conf_f"
+echo '#notify_lastmsg_num: Seconds before final normal notification expires (0=never)' >> "$conf_f"
+echo '#notify_errors_bool: Include possible errors in notifications' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Main Settings #' >> "$conf_f"
+echo '#main_ignorePackages_str: list of packages to ignore separated by spaces (in addition to pacman.conf)' >> "$conf_f"
+echo '#main_logdir_str: path to the log directory' >> "$conf_f"
+echo '#main_mirrorCountry_str: Countries separated by commas from which to pull updates. Default is automatic (geoip)' >> "$conf_f"
+echo '#main_testSite_str: url (without protocol) used to test internet connection' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Update Settings #' >> "$conf_f"
+echo '#update_downgrades_bool: Directs pacman to downgrade package if remote is older than local' >> "$conf_f"
+echo '#update_keys_bool: Check for security signature/key updates' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '# Custom Makepkg Flags for AUR packages (requires pikaur)' >> "$conf_f"
+echo '#zflag:packagename1,packagename2=--flag1,--flag2,--flag3' >> "$conf_f"
+echo '#' >> "$conf_f"
+echo '#' >> "$conf_f"
 DEFAULTIFS=$IFS; IFS=$'\n'
 for i in $(sort <<< "${!conf_a[*]}"); do
-	echo "$i=${conf_a[$i]}" | sudo tee -a "$conf_f"
+	echo "$i=${conf_a[$i]}" >> "$conf_f"
 done; IFS=$DEFAULTIFS
 }
 
-#--------------------------
-
-trouble(){ (echo;echo "#XS# `date` - $@";echo) |tee -a $log_f; }
-
-pacclean(){ 
-if [ ! "${conf_a[str_cleanLevel]}" = "off" ]; then
-    trouble "Cleaning package cache..."
-    if [ -d /var/cache/apacman/pkg ]; then rm -rf /var/cache/apacman/pkg/*; fi
-    if [ -d /var/cache/pikaur/pkg ]; then rm -rf /var/cache/pikaur/pkg/*; fi
-    [[ "${conf_a[str_cleanLevel]}" = "low" ]]  && paccache -rvk2
-    if [ "${conf_a[str_cleanLevel]}" = "high" ]; then
-        paccache -rvk0; [[ -d /var/cache/pikaur ]] && rm -rf /var/cache/pikaur/*; fi
-fi; }
-
 #Notification Functions
 
-killmsg(){ if [ "${conf_a[bool_notifyMe]}" = "$ctrue" ]; then killall xfce4-notifyd; fi; }
+killmsg(){ if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ]; then killall xfce4-notifyd; fi; }
 iconnormal(){ icon=ElectrodeXS; }
 iconwarn(){ icon=important; }
 iconcritical(){ icon=system-shutdown; }
 
 sendmsg(){
-    if [ "${conf_a[bool_notifyMe]}" = "$ctrue" ]; then
+    if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ]; then
         DISPLAY=$2 su $1 -c "dbus-launch notify-send -i $icon XS-AutoUpdate -u critical \"$3\"" & fi
 }
 
 sendall(){
-    if [ "${conf_a[bool_notifyMe]}" = "$ctrue" ]; then
+    if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ]; then
         getsessions; i=0; while [ $i -lt ${#s_usr[@]} ]; do
             sendmsg "${s_usr[$i]}" "${s_disp[$i]}" "$1"; i=$(($i+1))
         done; unset i
@@ -152,25 +150,140 @@ userlogon(){
 }
 
 
-#---Init Main---
+#---Init Config---
 
-#Define Log file
-[[ "${conf_a[str_log_d]}" = "" ]] && conf_a[str_log_d]="/var/log/xs"
-log_d="${conf_a[str_log_d]}"; log_f="${log_d}/auto-update.log"
+#Init Defaults
+
+typeset -A flag_a
+
+typeset -A conf_a; conf_a=(
+    [aur_1helper_str]="auto"
+    [aur_devel_bool]=$ctrue
+    [cln_1enable_bool]=$ctrue
+    [cln_aurpkg_bool]=$ctrue
+    [cln_aurbuild_bool]=$ctrue
+    [cln_orphan_bool]=$ctrue
+    [cln_paccache_num]=0
+    [flatpak_1enable_bool]=$ctrue
+    [notify_1enable_bool]=$ctrue
+    [notify_lastmsg_num]=20
+    [notify_errors_bool]=$ctrue
+    [main_ignorepkgs_str]=""
+    [main_logdir_str]="/var/log/xs"
+    [main_country_str]=""
+    [main_testsite_str]="www.google.com"
+    [update_downgrades_bool]=$ctrue
+    [update_keys_bool]=$ctrue
+    #legacy
+    [bool_detectErrors]=""
+    [bool_Downgrades]=""
+    [bool_notifyMe]=""
+    [bool_updateFlatpak]=""
+    [bool_updateKeys]=""
+    [str_cleanLevel]=""
+    [str_ignorePackages]=""
+    [str_log_d]=""
+    [str_mirrorCountry]=""
+    [str_testSite]=""
+)
+
+shopt -s extglob # needed for validconf
+validconf=@($(echo "${!conf_a[*]}"|sed "s/ /|/g"))
+
+conf_int0="notify_lastmsg_num"
+conf_intn1="cln_paccache_num"
+conf_legacy="bool_detectErrors bool_Downgrades bool_notifyMe bool_updateFlatpak bool_updateKeys str_cleanLevel str_ignorePackages str_log_d str_mirrorCountry str_testSite"
+
+#Load external config
+#Basic config validation
+
+if [ -f $conf_f ]; then
+    while read line; do
+        line=$(echo "$line" | cut -d ';' -f 1 | cut -d '#' -f 1)
+        if echo "$line" | grep -F '=' &>/dev/null; then
+            varname=$(echo "$line" | cut -d '=' -f 1)
+            case $varname in
+                $validconf) ;;
+                *) echo "$varname"|grep -F "zflag:" >/dev/null || continue
+            esac
+            line=$(echo "$line" | cut -d '=' -f 2-)
+            if [[ ! "$line" = "" ]]; then
+                #validate boolean
+                echo "$varname" | grep -F "bool" >/dev/null && if [[ ! ( "$line" = "$ctrue" || \
+                    "$line" = "$cfalse" ) ]]; then continue; fi
+                #validate numbers
+                if echo "$varname" | grep "num" >/dev/null; then 
+                    if [[ ! "$line" = "0" ]]; then let "line += 0"
+                    [[ "$line" = "0" ]] && continue; fi; fi
+                #validate integers 0+
+                if echo "$conf_int0" | grep "$varname" >/dev/null; then 
+                    if [[ "$line" -lt "0" ]]; then continue; fi; fi
+                #validate integers -1+
+                if echo "$conf_intn1" | grep "$varname" >/dev/null; then 
+                    if [[ "$line" -lt "-1" ]]; then continue; fi; fi
+                #validate aur_helper_str
+                if [[ "$varname" = "aur_helper_str" ]]; then case "$line" in
+                        auto|none|all|pikaur|apacman) ;;
+                        *) continue
+                esac; fi
+
+                conf_a[$varname]=$line
+                echo "$varname" | grep -F "zflag:" >/dev/null && \
+                    flag_a[$(echo "$varname" | cut -d ':' -f 2)]="${conf_a[$varname]}"
+
+            fi
+        fi
+    done < "$conf_f"; unset line; unset varname
+fi
+unset validconf; shopt -u extglob
+
+#Convert legacy settings
+
+case "${conf_a[str_cleanLevel]}" in
+    high) conf_a[cln_aurpkg_bool]="$ctrue";  conf_a[cln_aurbuild_bool]="$ctrue";  conf_a[cln_paccache_num]=0 ;;
+    low)  conf_a[cln_aurpkg_bool]="$cfalse"; conf_a[cln_aurbuild_bool]="$cfalse"; conf_a[cln_paccache_num]=2 ;;
+    off)  conf_a[cln_aurpkg_bool]="$cfalse"; conf_a[cln_aurbuild_bool]="$cfalse"; conf_a[cln_paccache_num]=-1
+esac
+
+[[ ! "${conf_a[bool_detectErrors]}" = "" ]]  && conf_a[notify_errors_bool]="${conf_a[bool_detectErrors]}"
+[[ ! "${conf_a[bool_Downgrades]}" = "" ]]    && conf_a[update_downgrades_bool]="${conf_a[bool_Downgrades]}"
+[[ ! "${conf_a[bool_notifyMe]}" = "" ]]      && conf_a[notify_1enable_bool]="${conf_a[bool_notifyMe]}"
+[[ ! "${conf_a[bool_updateFlatpak]}" = "" ]] && conf_a[flatpak_1enable_bool]="${conf_a[bool_updateFlatpak]}"
+[[ ! "${conf_a[bool_updateKeys]}" = "" ]]    && conf_a[update_keys_bool]="${conf_a[bool_updateKeys]}"
+[[ ! "${conf_a[str_ignorePackages]}" = "" ]] && conf_a[main_ignorepkgs_str]="${conf_a[str_ignorePackages]}"
+[[ ! "${conf_a[str_log_d]}" = "" ]]          && conf_a[main_logdir_str]="${conf_a[str_log_d]}"
+[[ ! "${conf_a[str_mirrorCountry]}" = "" ]]  && conf_a[main_country_str]="${conf_a[str_mirrorCountry]}"
+[[ ! "${conf_a[str_testSite]}" = "" ]]       && conf_a[main_testsite_str]="${conf_a[str_testSite]}"
+
+DEFAULTIFS=$IFS; IFS=$' '
+for i in $(sort <<< "$conf_legacy"); do
+	unset conf_a[$i]
+done; IFS=$DEFAULTIFS
+
+
+log_d="${conf_a[main_logdir_str]}"; log_f="${log_d}/auto-update.log"
+
+
+#---Main---
+
 
 #Start Sub-processes
 if [ "$1" = "backnotify" ]; then backgroundnotify; exit 0; fi
 if [ "$1" = "userlogon" ]; then userlogon; exit 0; fi
 
 #Init log dir, check for other running instances, start notifier
-mkdir -p "$log_d"; if [ ! -f "$log_f" ]; then echo "init">$log_f; fi
+mkdir -p "${conf_a[main_logdir_str]}"; if [ ! -d "${conf_a[main_logdir_str]}" ]; then conf_a[main_logdir_str]="/var/log/xs"; fi
+mkdir -p "${conf_a[main_logdir_str]}"; if [ ! -d "${conf_a[main_logdir_str]}" ]; then
+    echo "Critical error: could not create log directory"; sleep 10; exit; fi
+if [ ! -f "$log_f" ]; then echo "init">$log_f; fi
 if pidof -o %PPID -x "`basename "$0"`">/dev/null; then exit 0; fi #Only 1 main instance allowed
+exportconf
 if [ $# -eq 0 ]; then echo "`date` - XS-Update $vsndsp starting..." |tee $log_f; "$0" "XS"& exit 0; fi #Run in background
 
 #Wait up to 5 minutes for network
 trouble "Waiting for network..."
 waiting=1;waited=0; while [ $waiting = 1 ]; do
-    ping -c 1 "${conf_a[str_testSite]}" >/dev/null && waiting=0
+    ping -c 1 "${conf_a[main_testsite_str]}" >/dev/null && waiting=0
     if [ $waiting = 1 ]; then
         if [ $waited -ge 60 ]; then exit; fi
         sleep 5; waited=$(($waited+1))
@@ -198,10 +311,10 @@ if [ -f /var/lib/pacman/db.lck ]; then rm -f /var/lib/pacman/db.lck; fi
 #init main script and background notifications
 trouble "Init vars and notifier..."
 pacmirArgs="--geoip"
-[[ "${conf_a[str_mirrorCountry]}" = "" ]] || pacmirArgs="-c ${conf_a[str_mirrorCountry]}"
-[[ "${conf_a[str_ignorePackages]}" = "" ]] || pacignore="--ignore ${conf_a[str_ignorePackages]}"
-[[ "${conf_a[bool_Downgrades]}" = "$ctrue" ]] && pacdown=u
-[[ "${conf_a[aur_devel_bool]}" = "$ctrue" ]] && devel=--devel
+[[ "${conf_a[main_country_str]}" = "" ]] || pacmirArgs="-c ${conf_a[main_country_str]}"
+[[ "${conf_a[main_ignorepkgs_str]}" = "" ]] || pacignore="--ignore ${conf_a[main_ignorepkgs_str]}"
+[[ "${conf_a[update_downgrades_bool]}" = "$ctrue" ]] && pacdown="u"
+[[ "${conf_a[aur_devel_bool]}" = "$ctrue" ]] && devel="--devel"
 getsessions; i=0; while [ $i -lt ${#s_usr[@]} ]; do
 if [ -d "${s_home[$i]}/.cache" ]; then
     mkdir -p "${s_home[$i]}/.cache/xs"; echo "tmp" > "${s_home[$i]}/.cache/xs/logonnotify"
@@ -216,7 +329,7 @@ pacman-mirrors $pacmirArgs 2>&1 |sed 's/\x1B\[[0-9;]\+[A-Za-z]//g' |tr -cd '\11\
 trouble "Updating key packages..."
 pacman -S --needed --noconfirm archlinux-keyring manjaro-keyring manjaro-system 2>&1 |tee -a $log_f
 
-if [[ "${conf_a[bool_updateKeys]}" = "$ctrue" ]]; then
+if [[ "${conf_a[update_keys_bool]}" = "$ctrue" ]]; then
     trouble "Refreshing keys..."; (pacman-key --refresh-keys; sync;)  2>&1 |tee -a $log_f; fi
 
 trouble "Updating packages from main repos..."
@@ -276,26 +389,30 @@ if [[ "$use_apacman" = "1" ]]; then
 fi
 
 #Remove orphan packages, cleanup
-trouble "Removing orphan packages..."
-pacman -Rnsc $(pacman -Qtdq) --noconfirm  2>&1 |tee -a $log_f
+if [[ "${conf_a[cln_1enable_bool]}" = "$ctrue" ]]; then if [[ "${conf_a[cln_orphan_bool]}" = "$ctrue" ]]; then
+    if [[ ! "$(pacman -Qtdq)" = "" ]]; then
+        trouble "Removing orphan packages..."
+        pacman -Rnsc $(pacman -Qtdq) --noconfirm  2>&1 |tee -a $log_f
+    fi
+fi; fi
 pacclean
 
 #Update Flatpak
-if [[ "${conf_a[bool_updateFlatpak]}" = "$ctrue" ]]; then if type flatpak >/dev/null 2>&1; then
+if [[ "${conf_a[flatpak_1enable_bool]}" = "$ctrue" ]]; then if type flatpak >/dev/null 2>&1; then
     trouble "Updating flatpak..."
     flatpak update -y  2>&1 |tee -a $log_f
 fi; fi
 
 #Finish
 trouble "Update completed, final notifications and cleanup..."
-kill $bkntfypid; exportconf
-msg="System update finished"; grep "Total Installed Size:" $log_f && msg="$msg \nPackages successfully updated"
-grep "new signatures:" $log_f && msg="$msg \nSecurity signatures updated"
-grep "Total Removed Size:" $log_f && msg="$msg \nObsolete packages removed"
-if [ "${conf_a[bool_detectErrors]}" = "$ctrue" ]; then grep "error: failed " $log_f && msg="$msg \nSome packages encountered errors"; fi
+kill $bkntfypid
+msg="System update finished"; grep "Total Installed Size:" $log_f >/dev/null && msg="$msg \nPackages successfully updated"
+grep "new signatures:" $log_f >/dev/null && msg="$msg \nSecurity signatures updated"
+grep "Total Removed Size:" $log_f >/dev/null && msg="$msg \nObsolete packages removed"
+if [ "${conf_a[notify_errors_bool]}" = "$ctrue" ]; then grep "error: failed " $log_f >/dev/null && msg="$msg \nSome packages encountered errors"; fi
 if [ ! "$msg" = "System update finished" ]; then msg="$msg \nDetails: $log_f"; fi
 if [ "$msg" = "System update finished" ]; then msg="System up-to-date, no changes made"; fi
-normcrit=norm; grep "upgrading " $log_f |grep -v "tor-browser"|grep -E "linux[0-9]{2,3}" && normcrit=crit
+normcrit=norm; grep "upgrading " $log_f |grep -v "tor-browser"|grep -E "linux[0-9]{2,3}" >/dev/null && normcrit=crit
 [[ "$normcrit" = "norm" ]] && finalmsg_normal; [[ "$normcrit" = "crit" ]] && finalmsg_critical
 trouble "XS-done"; sleep 2; disown -a; sleep 2; systemctl stop xs-autoupdate.service >/dev/null 2>&1; exit 0
